refactor: split main package
