+++ cli/config.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"github.com/odpf/salt/config"
