+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/odpf/entropy/cli"
+
+func main() {
+	cli.Execute()
+}
